refactor: email container 분리
