+++ packages/gdsc-dju-admin/src/components/organisms/EmailContainer/index.tsx
@@ -1,0 +1,109 @@
+import { useAtom } from 'jotai';
+import React, { useRef } from 'react';
+import { useModalHandle } from '../../../hooks/useModalHandle';
+import {
+  EmailButtonWrapper,
+  InformationHeader,
+} from '../../../pages/Applicants/styled';
+import { TemplateSelectWrapper } from '../../../pages/EmailLog/styled';
+import { alertAtom } from '../../../store/alertAtom';
+import { templateAtom } from '../../../store/templateAtom';
+import { IApplicantTypeWithID } from '../../../types/applicant';
+import { GDSCButton } from '../../atoms/Button';
+import { TextInput } from '../../atoms/TextInput';
+import CheckBoxCard from '../../molecules/CheckBoxCard';
+import {
+  CheckboxSection,
+  CheckboxWrapper,
+  EmailContainerWrapper,
+  TemplateEmailWrapper,
+  TemplateText,
+} from './styled';
+
+interface Props {
+  filteredApplicants: IApplicantTypeWithID[];
+  checkedApplicants: Set<string>;
+  checkAllHandler: (isChecked: boolean) => void;
+  isAllChecked: boolean;
+}
+
+const EmailContainer = ({
+  checkedApplicants,
+  checkAllHandler,
+  isAllChecked,
+  filteredApplicants,
+}: Props) => {
+  const [alert, setAlert] = useAtom(alertAtom);
+  const [template, setTemplate] = useAtom(templateAtom);
+
+  const { openModal } = useModalHandle('EMAIL');
+  const templateRef = useRef<HTMLInputElement>(null);
+  const checkedApplicantHandler = (id: string, isChecked: boolean) => {
+    const newCheckedApplicants = new Set(checkedApplicants);
+    if (isChecked) {
+      newCheckedApplicants.add(id);
+    } else if (checkedApplicants.has(id)) {
+      newCheckedApplicants.delete(id);
+    }
+  };
+  const applyButtonHandler = () =>
+    checkedApplicants.size > 0
+      ? openModal()
+      : setAlert({
+          ...alert,
+          alertHandle: true,
+          alertMessage: '선택된 지원자가 없습니다.',
+        });
+  return (
+    <EmailContainerWrapper>
+      <InformationHeader>
+        <TemplateSelectWrapper>
+          {template && <TemplateText>선택한 템플릿: {template}</TemplateText>}
+          <TemplateEmailWrapper>
+            <TextInput
+              ref={templateRef}
+              placeholder={'템플릿을 입력해주세요.'}
+            />
+          </TemplateEmailWrapper>
+          <GDSCButton
+            color={'blue900'}
+            text={'템플릿 선택'}
+            onClick={() => setTemplate(templateRef.current?.value ?? '')}
+          />
+        </TemplateSelectWrapper>
+        <EmailButtonWrapper>
+          <GDSCButton
+            color={!isAllChecked ? 'blue200' : 'blue900'}
+            text={!isAllChecked ? '모두 선택' : '모두 해제'}
+            onClick={() => checkAllHandler(!isAllChecked)}
+            type={'button'}
+          />
+          <GDSCButton
+            color={'blue900'}
+            text={'이메일 전송'}
+            onClick={applyButtonHandler}
+            type={'button'}
+          />
+        </EmailButtonWrapper>
+      </InformationHeader>
+      {filteredApplicants && (
+        <CheckboxSection>
+          {filteredApplicants.map((applicant) => (
+            <CheckboxWrapper
+              key={applicant.id}
+              onDoubleClick={() => openModal(applicant.id)}
+            >
+              <CheckBoxCard
+                {...applicant}
+                checkedList={checkedApplicants}
+                setCheckedList={checkedApplicantHandler}
+              />
+            </CheckboxWrapper>
+          ))}
+        </CheckboxSection>
+      )}
+    </EmailContainerWrapper>
+  );
+};
+
+export default EmailContainer;
