+++ packages/gdsc-dju-admin/src/pages/Email/index.tsx
@@ -2,39 +2,24 @@ import emailjs from '@emailjs/browser';
 import { addDoc, collection } from 'firebase/firestore';
 import { AnimatePresence } from 'framer-motion';
 import { useAtom } from 'jotai';
-import React, { useEffect, useRef, useState } from 'react';
-import { GDSCButton } from '../../components/atoms/Button';
-import { TextInput } from '../../components/atoms/TextInput';
-import CheckBoxCard from '../../components/molecules/CheckBoxCard';
+import React, { useEffect, useState } from 'react';
 import AdminEmailCheckModal from '../../components/molecules/modal/AdminEmailCheckModal';
 import ApplicantModal from '../../components/molecules/modal/ApplicantModal';
+import EmailContainer from '../../components/organisms/EmailContainer';
+import SelectedEmailContainer from '../../components/organisms/SelectedEmailContainer';
+import { isDevelop } from '../../context/recruitInfo';
 import { db } from '../../firebase/firebase';
 import { useModalHandle } from '../../hooks/useModalHandle';
-import { isDevelop } from '../../pageData/recruitInfo';
 
 import { alertAtom } from '../../store/alertAtom';
 import { loaderAtom } from '../../store/loaderAtom';
+import { templateAtom } from '../../store/templateAtom';
 import { userAtom } from '../../store/userAtom';
 import { EmailLogType, IApplicantTypeWithID } from '../../types/applicant';
 import { getApplicants } from '../../utils/applicantsHandler';
-import {
-  AdminSectionWrapper,
-  EmailButtonWrapper,
-  InformationHeader,
-} from '../Applicants/styled';
-import { TemplateSelectWrapper } from '../EmailLog/styled';
-import {
-  CheckboxSection,
-  CheckboxWrapper,
-  EmailCategory,
-  EmailLeftInner,
-  EmailLeftWrapper,
-  EmailRightInner,
-  EmailRightWrapper,
-  SelectedBoxSection,
-  TemplateEmailWrapper,
-  TemplateText,
-} from './styled';
+import { AdminSectionWrapper } from '../Applicants/styled';
+import { AdminContainerInner } from '../styled';
+import { EmailLeftWrapper, EmailRightWrapper } from './styled';
 
 const Email = () => {
   const [alert, setAlert] = useAtom(alertAtom);
@@ -43,26 +28,18 @@ const Email = () => {
 
   const [filteredApplicants, setFilteredApplicants] =
     useState<IApplicantTypeWithID[]>();
-  const [checkedApplicants, setCheckedApplicants] = useState(new Set());
+  const [checkedApplicants, setCheckedApplicants] = useState<Set<string>>(
+    new Set(),
+  );
 
-  const [template, setTemplate] = useState<string | null>(null);
-  const templateRef = useRef<HTMLInputElement>(null);
-  const { openModal, closeModal } = useModalHandle('EMAIL');
+  const [template] = useAtom(templateAtom);
 
-  const checkedApplicantHandler = (id: string, isChecked: boolean) => {
-    const newCheckedApplicants = new Set(checkedApplicants);
-    if (isChecked) {
-      newCheckedApplicants.add(id);
-      setCheckedApplicants(newCheckedApplicants);
-    } else if (!isChecked && checkedApplicants.has(id)) {
-      newCheckedApplicants.delete(id);
-      setCheckedApplicants(newCheckedApplicants);
-    }
-  };
+  const { closeModal } = useModalHandle('EMAIL');
+
   const isAllChecked = checkedApplicants.size === filteredApplicants?.length;
 
-  const checkAllHandler = (isChecked: boolean) => {
-    if (isChecked) {
+  const checkAllHandler = () => {
+    if (isAllChecked) {
       setCheckedApplicants(new Set(filteredApplicants?.map((data) => data.id)));
     } else {
       setCheckedApplicants(new Set());
@@ -112,7 +89,6 @@ const Email = () => {
             email: applicant.email,
             name: applicant.name,
           });
-
           //로그 생성
           const emailLog: EmailLogType = {
             email: applicant.email,
@@ -144,15 +120,6 @@ const Email = () => {
     });
   };
 
-  const applyButtonHandler = () =>
-    checkedApplicants.size > 0
-      ? openModal()
-      : setAlert({
-          ...alert,
-          alertHandle: true,
-          alertMessage: '선택된 지원자가 없습니다.',
-        });
-
   useEffect(() => {
     (async function () {
       const applicants = await getApplicants(null);
@@ -165,85 +132,35 @@ const Email = () => {
   });
 
   return (
-    <AnimatePresence>
-      <AdminSectionWrapper>
-        <ApplicantModal />
-        {selectApplicants && (
-          <AdminEmailCheckModal
-            applicants={selectApplicants}
-            sendEmail={sendEmailHandler}
-            template={template}
-          />
-        )}
-
-        <EmailLeftWrapper>
-          <EmailLeftInner>
-            <EmailCategory>선택한 이메일</EmailCategory>
+    <AdminContainerInner>
+      <AnimatePresence>
+        <AdminSectionWrapper>
+          <ApplicantModal />
+          {selectApplicants && (
+            <AdminEmailCheckModal
+              applicants={selectApplicants}
+              sendEmail={sendEmailHandler}
+              template={template}
+            />
+          )}
+          <EmailLeftWrapper>
             {selectApplicants && (
-              <SelectedBoxSection>
-                {selectApplicants.map((applicant) => (
-                  <div key={`check-${applicant.id}`}>
-                    <CheckBoxCard {...applicant} disabled={true} />
-                  </div>
-                ))}
-              </SelectedBoxSection>
+              <SelectedEmailContainer selectApplicants={selectApplicants} />
             )}
-          </EmailLeftInner>
-        </EmailLeftWrapper>
-        <EmailRightWrapper>
-          <EmailRightInner>
-            <InformationHeader>
-              <TemplateSelectWrapper>
-                {template && (
-                  <TemplateText>선택한 템플릿: {template}</TemplateText>
-                )}
-                <TemplateEmailWrapper>
-                  <TextInput
-                    ref={templateRef}
-                    placeholder={'템플릿을 입력해주세요.'}
-                  />
-                </TemplateEmailWrapper>
-                <GDSCButton
-                  color={'blue900'}
-                  text={'템플릿 선택'}
-                  onClick={() => setTemplate(templateRef.current?.value ?? '')}
-                />
-              </TemplateSelectWrapper>
-              <EmailButtonWrapper>
-                <GDSCButton
-                  color={!isAllChecked ? 'blue200' : 'blue900'}
-                  text={!isAllChecked ? '모두 선택' : '모두 해제'}
-                  onClick={() => checkAllHandler(!isAllChecked)}
-                  type={'button'}
-                />
-                <GDSCButton
-                  color={'blue900'}
-                  text={'이메일 전송'}
-                  onClick={applyButtonHandler}
-                  type={'button'}
-                />
-              </EmailButtonWrapper>
-            </InformationHeader>
+          </EmailLeftWrapper>
+          <EmailRightWrapper>
             {filteredApplicants && (
-              <CheckboxSection>
-                {filteredApplicants.map((applicant) => (
-                  <CheckboxWrapper
-                    key={applicant.id}
-                    onDoubleClick={() => openModal(applicant.id)}
-                  >
-                    <CheckBoxCard
-                      {...applicant}
-                      checkedList={checkedApplicants}
-                      setCheckedList={checkedApplicantHandler}
-                    />
-                  </CheckboxWrapper>
-                ))}
-              </CheckboxSection>
+              <EmailContainer
+                checkedApplicants={checkedApplicants}
+                checkAllHandler={checkAllHandler}
+                filteredApplicants={filteredApplicants}
+                isAllChecked={isAllChecked}
+              />
             )}
-          </EmailRightInner>
-        </EmailRightWrapper>
-      </AdminSectionWrapper>
-    </AnimatePresence>
+          </EmailRightWrapper>
+        </AdminSectionWrapper>
+      </AnimatePresence>
+    </AdminContainerInner>
   );
 };
