+++ packages/gdsc-dju-admin/src/pages/EmailLog/styled.ts
@@ -10,6 +10,6 @@ export const TemplateSelectWrapper = styled.div`
   margin: 20px 0;
 `;
 export const LogWrapper = styled.div`
-  height: calc(100vh - 70px - 82px - 120px);
+  height: 100%;
   overflow: scroll;
 `;
