+++ packages/gdsc-dju-admin/src/pages/styled.ts
@@ -1,12 +1,12 @@
 import styled from 'styled-components';
 
 export const AdminContainerInner = styled.div`
-  background: ${({ theme }) => theme.colors.grey50};
   height: 100%;
   padding: 30px 30px;
   overflow: scroll;
 `;
 export const AdminContainer = styled.div`
+  background: ${({ theme }) => theme.colors.grey50};
   -webkit-box-flex: 1;
   flex-grow: 1;
   height: 100%;
@@ -19,9 +19,9 @@ export const AdminContainerWrapper = styled.div`
   overflow: hidden;
 `;
 export const SidebarContainer = styled.div`
-  width: 225px;
+  width: 200px;
   height: 100%;
-  border-right: 1px solid ${({ theme }) => theme.colors.grey300};
+  border-right: 1px solid ${({ theme }) => theme.colors.grey100};
   flex-shrink: 0;
   padding-top: 35px;
   position: relative;
