+++ packages/gdsc-dju-admin/src/store/templateAtom.ts
@@ -1,0 +1,3 @@
+import { atom } from 'jotai';
+
+export const templateAtom = atom<string | null>(null);
